Cannot do it like this
the task counter varible is not shared between my ranks, so it cannot be
used to keep track of how much work has been done. In threads, the main
thread is a default centralised thread, but this is not the case in
ranks. Either I keep a centralised rank that takes care of this stuff OR
I distribute the tasks and then call MasterFunctions that many times.

I'll decide what is worth doing tomorrow.